+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Nomenclature :
 
 ### Unreleased
 
+### [0.1.10] - 2023-04-17
+
+#### Updated
+
+- Updated and moved balance-card to a common component and used it in owner detail and tenant detail views.
+
 ### [0.1.9] - 2023-04-16
 
 #### Added